pyproject.toml
[project]
name = "pbianalyzer"
dynamic = ["version"]
description = "Power BI on Databricks SQL Health Check Optimiser"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "python-multipart>=0.0.9",
    "pbixray>=0.5.0",
    "fpdf2>=2.8.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.12",
]

[tool.apx.metadata]
app-name = "pbianalyzer"
app-slug = "pbianalyzer"
app-entrypoint = "pbianalyzer.backend.app:app"
api-prefix = "/api"
metadata-path = "src/pbianalyzer/_metadata.py"

[tool.apx.ui]
root = "src/pbianalyzer/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.version]
path = "src/pbianalyzer/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/pbianalyzer/__dist__", "src/pbianalyzer/_metadata.py"]
exclude = ["src/pbianalyzer/ui"]
